basic flask working and coming through to dev machine
