+++ requirements.txt
@@ -1,4 +1,4 @@
-python-can
-AWSIoTPythonSDK
-pytz
-
+python-can~=3.3.4
+AWSIoTPythonSDK~=1.4.9
+pytz~=2021.3
+Flask~=2.0.2
